+++ source/qscript/misc.d
@@ -7,7 +7,7 @@ import std.conv:to;
 
 string parseStr(string s){
 	string r;
-	for (uint i=0;i<s.length;i++){
+	for (ulong i=0;i<s.length;i++){
 		if (s[i]=='\\'){
 			i++;
 			switch (s[i]){
@@ -49,7 +49,7 @@ T max (T)(T[] dat){
 T[] convArrayType(T,T2)(T2[] dat){
 	T[] r;
 	r= new T[dat.length];
-	for (uint i=0;i<dat.length;i++){
+	for (ulong i=0;i<dat.length;i++){
 		r[i]=to!T(dat[i]);
 	}
 	return r;
@@ -59,7 +59,7 @@ string[] fileToArray(string fname){
 	File f = File(fname,"r");
 	string[] r;
 	string line;
-	int i=0;
+	long i=0;
 	r.length=0;
 	while (!f.eof()){
 		if (i+1>=r.length){
@@ -77,14 +77,14 @@ string[] fileToArray(string fname){
 	return r;
 }
 
-T[] del(T)(T[] dat, uint pos, uint count=1){
+T[] del(T)(T[] dat, ulong pos, ulong count=1){
 	T[] ar1, ar2;
 	ar1 = dat[0..pos];
 	ar2 = dat[pos+count..dat.length];
 	return ar1~ar2;
 }
 
-T[] insert(T)(T[] dat, T[] ins, uint pos){
+T[] insert(T)(T[] dat, T[] ins, ulong pos){
 	T[] ar1, ar2;
 	ar1 = dat[0..pos];
 	ar2 = dat[pos..dat.length];
